Fix explosion close to the left border

+++ src/explosion.h
@@ -8,13 +8,13 @@
 
 struct explosion_t
 {
-    u8 x;
-    u8 y;
-    u8 frame;
-    u8 pad;
+    u16 x;
+    u16 y;
+    u8  frame;
+    u8  pad;
 };
 
-void init_explosion(u8 id, u8 x, u8 y);
+void init_explosion(u8 id, u16 x, u16 y);
 
 u8 update_explosion(u8 id);
